Debugging and organizing Evaluator folder
